prevent compiler warning

+++ lib/sslgen.c
@@ -18,7 +18,7 @@
  * This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY
  * KIND, either express or implied.
  *
- * $Id: sslgen.c,v 1.2 2005-04-08 09:25:48 bagder Exp $
+ * $Id: sslgen.c,v 1.3 2005-04-19 23:38:57 bagder Exp $
  ***************************************************************************/
 
 /* This file is for "generic" SSL functions that all libcurl internals should
@@ -481,7 +481,7 @@ int Curl_ssl_recv(struct connectdata *conn, /* connection data */
       return -1;
   }
 
-  return nread;
+  return (int)nread;
 
 #else /* USE_SSL */
   (void)conn;
